Module-level refactoring and public API commenting
