+++ src/TesApi.Web/GroupableSet.cs
@@ -4,76 +4,9 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
-using Tes.Repository;
 
 namespace TesApi.Web
 {
-    /// <summary>
-    /// TODO
-    /// </summary>
-    /// <typeparam name="T"></typeparam>
-    public interface IHasRepositoryItem<T> where T : RepositoryItem<T>
-    {
-        /// <summary>
-        /// Returns the associated <see cref="RepositoryItem{T}"/>.
-        /// </summary>
-        T RepositoryItem { get; }
-
-        /// <summary>
-        /// TODO
-        /// </summary>
-        /// <param name="replacementItem"></param>
-        void ReplaceRepositoryItem(T replacementItem);
-
-        /// <summary>
-        /// TODO
-        /// </summary>
-        IHasChangesRepositoryItem<T> ChangesRepositoryItem { get; }
-    }
-
-    /// <summary>
-    /// TODO
-    /// </summary>
-    public interface IHasChangesRepositoryItem<T> where T : RepositoryItem<T>
-    {
-        /// <summary>
-        /// TODO
-        /// </summary>
-        bool HasChanges { get; }
-
-        /// <summary>
-        /// TODO
-        /// </summary>
-        void ResetChanges();
-    }
-
-    /// <inheritdoc/>
-    public class HasChangesRepositoryItem<T> : IHasChangesRepositoryItem<T> where T : RepositoryItem<T>
-    {
-        private readonly IHasRepositoryItem<T> _repositoryItemOwner;
-        private int _checkpoint;
-
-        /// <summary>
-        /// TODO
-        /// </summary>
-        /// <param name="repositoryItemOwner"></param>
-        /// <exception cref="ArgumentNullException"></exception>
-        public HasChangesRepositoryItem(IHasRepositoryItem<T> repositoryItemOwner)
-        {
-            _repositoryItemOwner = repositoryItemOwner ?? throw new ArgumentNullException(nameof(repositoryItemOwner));
-            ResetChanges();
-        }
-
-        /// <inheritdoc/>
-        public bool HasChanges
-            => _checkpoint != _repositoryItemOwner.RepositoryItem.GetHashCode();
-
-        /// <inheritdoc/>
-        public void ResetChanges()
-            => _checkpoint = _repositoryItemOwner.RepositoryItem.GetHashCode();
-    }
-
     /// <summary>
     /// Represents a set of values.
     /// </summary>
@@ -251,68 +184,6 @@ namespace TesApi.Web
         /// <inheritdoc/>
         IEnumerator IEnumerable.GetEnumerator()
             => GetEnumerator();
-        #endregion
-    }
-
-    /// <summary>
-    /// Represents a set of values.
-    /// </summary>
-    /// <typeparam name="TElement">The type of elements in the hash set.</typeparam>
-    /// <typeparam name="TRepositoryItem"></typeparam>
-    public class GroupablSetWithRepositoryElements<TElement, TRepositoryItem> : GroupableSet<TElement>, ISet<TElement> where TElement : IHasRepositoryItem<TRepositoryItem> where TRepositoryItem : RepositoryItem<TRepositoryItem>
-    {
-        #region Constructors
-        /// <summary>
-        /// Initializes a new instance of the <see cref="GroupablSetWithRepositoryElements{TElement, TRepositoryItem}"/> class that is empty and uses the default equality comparer for the set type.
-        /// </summary>
-        public GroupablSetWithRepositoryElements() : base() { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="GroupablSetWithRepositoryElements{TElement, TRepositoryItem}"/> class that uses the default equality comparer for the set type, contains elements copied from the specified collection, and has sufficient capacity to accommodate the number of elements copied.
-        /// </summary>
-        /// <param name="collection">The collection whose elements are copied to the new set.</param>
-        /// <exception cref="ArgumentNullException">collection is null.</exception>
-        public GroupablSetWithRepositoryElements(IEnumerable<TElement> collection) : base(collection) { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="GroupablSetWithRepositoryElements{TElement, TRepositoryItem}"/> class that is empty and uses the specified equality comparer for the set type.
-        /// </summary>
-        /// <param name="comparer">The <see cref="IEqualityComparer{T}"/> implementation to use when comparing values in the set, or null to use the default <see cref="IEqualityComparer{T}"/> implementation for the set type.</param>
-        public GroupablSetWithRepositoryElements(IEqualityComparer<TElement> comparer) : base(comparer) { }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="GroupablSetWithRepositoryElements{TElement, TRepositoryItem}"/> class that uses the specified equality comparer for the set type, contains elements copied from the specified collection, and has sufficient capacity to accommodate the number of elements copied.
-        /// </summary>
-        /// <param name="collection">The collection whose elements are copied to the new set.</param>
-        /// <param name="comparer">The <see cref="IEqualityComparer{T}"/> implementation to use when comparing values in the set, or null to use the default <see cref="IEqualityComparer{T}"/> implementation for the set type.</param>
-        public GroupablSetWithRepositoryElements(IEnumerable<TElement> collection, IEqualityComparer<TElement> comparer) : base(collection, comparer) { }
-        #endregion
-
-        #region Public methods
-        /// <summary>
-        /// Enumerates the <see cref="RepositoryItem{T}"/> of each element in this object.
-        /// </summary>
-        /// <returns>An <see cref="IEnumerable{T}"/> of the <see cref="RepositoryItem{T}"/> of each element.</returns>
-        public IEnumerable<TRepositoryItem> GetRepositoryItems()
-            => this.Select(e => e.RepositoryItem);
-
-
-        /// <summary>
-        /// TODO
-        /// </summary>
-        public bool HasChanges
-            => this.Any(e => e.ChangesRepositoryItem.HasChanges);
-
-        /// <summary>
-        /// TODO
-        /// </summary>
-        public void ResetChanges()
-        {
-            foreach (var element in this)
-            {
-                element.ChangesRepositoryItem.ResetChanges();
-            }
-        }
         #endregion
     }
 }
